requirements.txt
attrs==24.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
outcome==1.3.0.post0
packaging==24.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
webdriver-manager==4.0.2
wsproto==1.2.0
zipp==3.15.0
